예외 처리 추가

+++ routes/posts.js
@@ -25,36 +25,44 @@ router.get('/:postId', async (req, res) => {
     .match({ postId: Number(postId) })
     .project(fields)
 
+  if (!post.length) {
+    return res.status(404).json({ message: `${postId}번 게시글이 존재하지 않습니다.` })
+  }
+
   res.json({ post })
 })
 
 router.post('', auth, async (req, res) => {
-  const { title, password, content } = req.body
-  const { nickname } = res.locals.user
+  try {
+    const { title, password, content } = req.body
+    const { nickname } = res.locals.user
 
-  const postsCount = await Counts.find({ countId: 'posts' })
+    const postsCount = await Counts.find({ countId: 'posts' })
 
-  if (!postsCount.length) {
-    await Counts.create({ countId: 'posts', counts: 0 })
+    if (!postsCount.length) {
+      await Counts.create({ countId: 'posts', counts: 0 })
+    }
+
+    await Counts.updateOne({ countId: 'posts' }, { $inc: { counts: 1 } })
+
+    const postId = (await Counts.find({ countId: 'posts' }))[0].counts
+    const postedDate = new Date()
+
+    const newPost = {
+      postId,
+      nickname,
+      title,
+      password,
+      content,
+      postedDate,
+    }
+
+    await Posts.create(newPost)
+
+    res.json(newPost)
+  } catch {
+    res.status(400).json({ message: '게시글 작성 양식이 올바르지 않습니다.' })
   }
-
-  await Counts.updateOne({ countId: 'posts' }, { $inc: { counts: 1 } })
-
-  const postId = (await Counts.find({ countId: 'posts' }))[0].counts
-  const postedDate = new Date()
-
-  const newPost = {
-    postId,
-    nickname,
-    title,
-    password,
-    content,
-    postedDate,
-  }
-
-  await Posts.create(newPost)
-
-  res.json(newPost)
 })
 
 router.patch('/:postId', auth, async (req, res) => {
@@ -62,7 +70,11 @@ router.patch('/:postId', auth, async (req, res) => {
   const { nickname } = res.locals.user
   const { title, password, content } = req.body
 
-  await Posts.updateOne({ postId, nickname, password }, { title, content })
+  const result = await Posts.updateOne({ postId, nickname, password }, { title, content })
+
+  if (!result.modifiedCount) {
+    return res.status(400).json({ message: '게시글 수정 양식이 올바르지 않습니다.' })
+  }
 
   const post = await Posts.findOne({ postId })
 
@@ -74,7 +86,15 @@ router.delete('/:postId', auth, async (req, res) => {
   const { nickname } = res.locals.user
   const { password } = req.body
 
-  await Posts.deleteOne({ postId, nickname, password })
+  if (!password) {
+    return res.status(400).json({ message: '비밀번호를 입력해주세요' })
+  }
+
+  const result = await Posts.deleteOne({ postId, nickname, password })
+
+  if (!result.deletedCount) {
+    return res.status(400).json({ message: '게시글 수정 양식이 올바르지 않습니다.' })
+  }
 
   res.json({ message: '삭제 완료' })
 })
@@ -84,6 +104,10 @@ router.patch('/:postId/likes', auth, async (req, res) => {
   const { nickname } = res.locals.user
 
   const post = await Posts.findOne({ postId })
+
+  if (!post) {
+    return res.status(404).json({ message: `${postId}번 게시글이 존재하지 않습니다.` })
+  }
 
   if (post.likes.includes(nickname)) {
     post.likes.pull(nickname)
